Add alternate font set
This adds a 'alternate' font set that can be swapped to with a keybind
(c-m-f by default). In addition to allowing people to have an alternate
font for certain applications, swapping to the alternate font, and then
back, causes the current font to be reloaded, which is quite useful when
you are setting up a new font to use.